Use Includes Only Where Needed
Remove unused includes in libblinkstick header (which would be included in client code) and instead only include it where it is needed (the *.c) file. As a consequence you have to now include the same files in the CLI but this makes it more explicit and portable. Also added include guard to libblinkstick and windows check for including <unistd.h> which doesn't exist by default on windows.