Updates for BlinkStick Pro
Update set color function to include input for the channel of the blinkstick. This is valid for all blinksticks but is usually ignored for non-pro blinksticks. Also changed the use of `hid_write()` to instead be `hid_send_feature_report()`. This may be a windows specific issue but `hid_write()` does not work with blinkstick pros on windows. Added a new function to allow setting the "mode" of the blinkstick (again valid for blinkstick pros). Added an enum that gives the possible modes that can be set. Minor other cleanups (use of const and free()). Updated the CLI to include input of the channel and use the new functions. Use of the mode from CLI has not been added.

Resolves #15